Bugfix for deleting cards/groups out of order not saving.
Added code in cardsort.js to removeGroup and removeCard functions to reapply id and name numbering in the forms. Form validation fails if numbering is not contiguous, which was preventing saving after removing cards or groups out of order (e.g. deleting the first card and leaving the rest, then attempting to save).